test before delete

+++ shell/push-coverage.sh
@@ -10,7 +10,7 @@ for lib_name in $lib_folders; do
     cover_folder="$lib_name/target/site/jacoco"
     cover_file="$lib_name/target/site/jacoco/index.html"
 
-    rm -R docs/$lib_name/
+    if [[ -d docs/$lib_name ]]; then rm -R docs/$lib_name/ ; fi
     cp -R $cover_folder docs/$lib_name/
 done
